+++ CommHandler/CHANGELOG.md
@@ -11,10 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - client for SGr declaration library
 - JSONata expressions as response query in REST / messaging data points and message filters
+- response or template query can be used in write operations
 
 ### Changed
 
-- requires updated driver implementations
+- requires updated driver implementations supporting driver API 2.4
 
 
 ## [2.4.2] - 2025-10-09